added route folder and controller folder
