+++ app.js
@@ -7,6 +7,14 @@ const PORT = 7000;
 const path = require("path");
 const dbConnection = require("./database/mongodb");
 const credit_card = path.join(__dirname, "frontend") ;
+app.use(bodyParser.json());
+//import routes 
+const adminRoutes = require('./routes/admin');
+//import routes from frontend
+const adminRoutesFrontend = require('./routes/frontend');
+
+app.use(adminRoutes);
+app.use(adminRoutesFrontend);
 
 // Allows you to recieve data from postman body
 app.use(express.json());
@@ -16,83 +24,25 @@ app.use(bodyParser.json());
 //APP.GET /ALLCARDS
 
 // Fetch data from Mongo through API
-app.get("/allCard", async (req, res) => {
-  let data = await dbConnection();
-  data = await data.find({}).toArray();
-  // Assume you have a MongoDB collection named "users"
-  res.send(data);
-});
+
 
 //GET /
 
-// Main page with html
-app.get("/", async (req, res) => {
-  res.sendFile(`${credit_card}/index.html`);
-});
 
 
-// // Luhn10 Check for credit card
-function isValidCreditCardNumber(creditCardNumber) {
-  //card number between 12-19 and string. 
-  if (!creditCardNumber || typeof creditCardNumber !== 'string' || creditCardNumber.length < 12 || creditCardNumber.length > 19) {
-    return false;
-  }
 
-  // Remove all non-digit characters from the credit card number
 
-  
-
-const digits = creditCardNumber.replace(/\D/g, '');
-
-  // Check if the remaining characters form a valid Luhn 10 number
-  const sum = Array.from(digits).reverse().reduce((acc, digit, i) => {
-    const value = parseInt(digit, 10);
-    return acc + (i % 2 === 0 ? value : (value < 5 ? value * 2 : (value * 2) - 9));
-  }, 0);
-
-  return sum % 10 === 0;
-}
 
 
 //POST /
-app.post("/", async (req, res) => {
-  try {
-    const creditCardNumber = req.body.card_number;
-    if (!isValidCreditCardNumber(creditCardNumber)) {
-      console.log(creditCardNumber);
-      res.send({ status: "INVALID CREDIT CARD -> Format should be 4444555566667777" });
-      console.log("invlid ceredit card")
-      return;
-    }
-  
-    let data = await dbConnection();
-    let addCard = await data.insertOne(req.body);
-  
-    console.log(req.body);
-    res.send(addCard);
 
-
-  }catch(e){
-    console.log(e);
-  }
-
-});
 
 //PUT /
 
-//Update ddb using id in postman body 
-app.put("/", async (req, res) => {
-  let data = await dbConnection();
-  let result = data.updateOne({ id: req.body.id }, { $set: req.body });
-  res.send({ status: "updated" });
-});
+
 
 //DELETE /:id
-app.delete("/:id", async (req, res) => {
-  const data = await dbConnection();
-  const result = await data.deleteOne({ _id: new mongodb.ObjectId(req.params.id)});
-  res.send({ status: "updated" });
-})
+
 //getAllCard();
 
 app.listen(PORT, () => console.log(`LIVE on Localhost on :${PORT}`));
